requirements.txt
# Auto-Grader — Python dependencies
# Install: pip install -r requirements.txt  (preferably inside a virtual environment)

pdf2image>=1.17.0
pytesseract>=0.3.13
numpy>=1.24.0
pikepdf>=10.0.0
Pillow>=10.0.0
google-genai>=1.0.0
openai>=1.0.0  # For Kimi API support
pydantic>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
